FIXES move staking pool data fetching into wizard page
